Add modal components for voting, soldout, and ingredients
