+++ src/components/molecules/ProductDescription.jsx
@@ -1,8 +1,8 @@
 import React, { Component } from 'react'
 import styled from 'styled-components'
-import { FaTimes } from 'react-icons/fa'
-
-import Modal from '../../layouts/Modal'
+import ModalLayout from '../../layouts/ModalLayout'
+import ModalVote from '../atoms/ModalVote'
+import ModalSoldOut from '../atoms/ModalSoldOut'
 import { StyledHr, StyledButton, QuantityAdjustButton } from '../atoms'
 
 const Container = styled.div`
@@ -43,30 +43,6 @@ const Container = styled.div`
     }
   }
 `
-//inline style for modal background (used inline style so we did not need to add more styles to top level index.css)
-//styled components do not apply due to use of portal
-const outerModalStyle = {
-  display: 'flex',
-  alignItems: 'center',
-  position: 'fixed',
-  top: 0,
-  bottom: 0,
-  left: 0,
-  right: 0,
-  width: '100vw',
-  height: '100vh',
-  backgroundColor: 'rgba(29, 29, 29, 0.7)',
-  zIndex: 100000,
-}
-
-const modalStyle = {
-  padding: '2em',
-  backgroundColor: 'white',
-  width: '70%',
-  height: 'fit-content',
-  margin: '0 auto',
-  maxWidth: '30rem',
-}
 
 class ProductDescription extends Component {
   constructor(props) {
@@ -128,13 +104,6 @@ class ProductDescription extends Component {
     }
   }
   render() {
-    let charities
-    if (this.props.charities !== null) {
-      charities = this.props.charities.charities.map(charity => (
-        <option value={charity}>{charity}</option>
-      ))
-    }
-
     const productCopy = this.props.productCopy.map((statement, i) => {
       if (statement.content.length === 1)
         return <p key={i}>{statement.content[0].value}</p>
@@ -165,7 +134,11 @@ class ProductDescription extends Component {
           />
           <StyledButton
             type="button"
-            onClick={charities ? this.toggleModal : this.handleSubmit}
+            onClick={
+              this.props.charities || !this.props.availableForSale
+                ? this.toggleModal
+                : this.handleSubmit
+            }
             height={'2.5rem'}
             width={'18rem'}
             fontSize={'.65rem'}
@@ -177,72 +150,20 @@ class ProductDescription extends Component {
         {this.state.hideModal ? (
           ''
         ) : (
-          //Modal component is rendered with a portal
-          //everything inside Modal is rendered at the top level (index.js)
-          <Modal>
-            <div style={outerModalStyle}>
-              <div id="modalBox" style={modalStyle}>
-                <div
-                  style={{
-                    display: 'flex',
-                    flex: 'wrap',
-                    flexDirection: 'row',
-                    justifyContent: 'space-between',
-                  }}
-                >
-                  <div>
-                    <p style={{ fontWeight: 'normal', fontSize: '20px' }}>
-                      Cast your vote:
-                    </p>
-                  </div>
-
-                  <div>
-                    <button
-                      style={{
-                        backgroundColor: 'none',
-                        border: 'none',
-                        outline: 0,
-                      }}
-                      onClick={this.toggleModal}
-                    >
-                      <FaTimes />
-                    </button>
-                  </div>
-                </div>
-
-                <div>
-                  <p>Our donation will be sent to the most popular charity.</p>
-                </div>
-                <div>
-                  <form action="" onSubmit={this.handleSubmit}>
-                    <select
-                      name="charities"
-                      id="charitiesDropDown"
-                      onChange={this.handleChangeModal}
-                      style={{
-                        minWidth: '14rem',
-                        marginBottom: '1rem',
-                        width: '97%',
-                      }}
-                    >
-                      <option value="" />
-                      {charities}
-                    </select>
-                    <div>
-                      <StyledButton
-                        width="14rem"
-                        margin="0"
-                        fontSize=".65rem"
-                        height="2.5rem"
-                      >
-                        <b>ADD TO BAG</b>
-                      </StyledButton>
-                    </div>
-                  </form>
-                </div>
-              </div>
-            </div>
-          </Modal>
+          //ModalLayout component is rendered with a portal
+          //everything inside ModalLayout is rendered at the top level (index.js)
+          <ModalLayout>
+            {this.props.availableForSale ? (
+              <ModalVote
+                charities={this.props.charities}
+                toggleModal={this.toggleModal}
+                handleChangeModal={this.handleChangeModal}
+                handleSubmit={this.handleSubmit}
+              />
+            ) : (
+              <ModalSoldOut toggleModal={this.toggleModal} />
+            )}
+          </ModalLayout>
         )}
       </Container>
     )
