+++ src/templates/ProductTemplate.js
@@ -15,6 +15,7 @@ class ProductTemplate extends Component {
   render() {
     const data = this.props.pageContext.node
     const charities = data.charities
+    const availableForSale = this.props.pageContext.availableForSale
     const images = data.images.map(img => img.file.url)
 
     let productDetails
@@ -35,6 +36,7 @@ class ProductTemplate extends Component {
             price={data.price}
             sku={data.sku}
             charities={charities}
+            availableForSale={availableForSale}
           />
         </Container>
         {productDetails}
